+++ src/components/Application/SectionSummary.tsx
