+++ src/components/index.ts
@@ -3,7 +3,7 @@ import {
   ApplicationHeader,
   ApplicationSelectType,
   ApplicationStart,
-  ApplicationSummary,
+  SectionSummary,
   ProgressBar,
 } from './Application'
 import Approval from './Approval'
@@ -35,7 +35,6 @@ export {
   ApplicationHeader,
   ApplicationSelectType,
   ApplicationStart,
-  ApplicationSummary,
   Approval,
   Admin,
   Config,
@@ -47,6 +46,7 @@ export {
   Notification,
   NotificationsList,
   NoMatch,
+  SectionSummary,
   ProgressBar,
   Product,
   ProductList,
